make ArenaKey 64-bit aligned by default
this allows hashing with rust_hash::FxHasher to be a load + mul, and
nothing else on x86(_64)!

Make the default generation `g32`, since exhausting a 32-bit integer on
a single slot is unlikely, this shouldn't cause any major issues.
And it will make it more likely that the above ^^^ hash opt will kick in